+++ src/components/NumberInput.js
@@ -10,7 +10,6 @@ export default function NumberInput(props = {
     name: "",
     max: "",
     min: "",
-    required: false,
     value: "",
 })
 {
@@ -39,7 +38,6 @@ export default function NumberInput(props = {
             min: props.min,
             value: props.value,
             class: styles["number-input"],
-            ...(props.required && { required: true }),
         },
         []
     );